requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2.5
django-cache-url==3.4.5
django-debug-toolbar==5.2.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
environs==14.2.0
geographiclib==2.1
geopy==2.4.1
idna==3.10
marshmallow==4.0.0
phonenumbers==9.0.13
pillow==11.2.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
